pyproject.toml
[project]
name = "ancestor-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datastar-py>=0.6.3",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "quart>=0.20.0",
]
